# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "CancelVideoJobsRequest": {
        "properties": {
          "jobIds": {
            "items": {
              "type": "string"
            },
            "title": "Jobids",
            "type": "array"
          }
        },
        "required": [
          "jobIds"
        ],
        "title": "CancelVideoJobsRequest",
        "type": "object"
      },
      "CancelledVideoJobsView": {
        "properties": {
          "cancelledJobIds": {
            "items": {
              "type": "string"
            },
            "title": "Cancelledjobids",
            "type": "array"
          }
        },
        "required": [
          "cancelledJobIds"
        ],
        "title": "CancelledVideoJobsView",
        "type": "object"
      },
...
      "GenerationSubmitRequest": {
        "properties": {
          "seed": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seed"
          }
        },
        "title": "GenerationSubmitRequest",
        "type": "object"
      },
      "GenerationSummary": {
        "properties": {
          "quality": {
            "title": "Quality",
            "type": "string"
          },
          "resolution": {
            "title": "Resolution",
            "type": "string"
          }
        },
        "required": [
          "resolution",
          "quality"
        ],
        "title": "GenerationSummary",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "JobView": {
        "properties": {
          "completedAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completedat"
          },
          "errorCode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Errorcode"
          },
          "errorMessage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Errormessage"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "providerJobId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Providerjobid"
          },
          "startedAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Startedat"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "submittedAt": {
            "format": "date-time",
            "title": "Submittedat",
            "type": "string"
          },
          "taskId": {
            "title": "Taskid",
            "type": "string"
          }
        },
        "required": [
          "id",
          "taskId",
          "status",
          "providerJobId",
          "submittedAt",
          "startedAt",
          "completedAt",
          "errorCode",
          "errorMessage"
        ],
        "title": "JobView",
        "type": "object"
      },
...
      "ProjectCoverSelection": {
        "properties": {
          "assetId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assetid"
          },
          "kind": {
            "enum": [
              "auto",
              "asset",
              "video"
            ],
            "title": "Kind",
            "type": "string"
          },
          "resultId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resultid"
          },
          "seconds": {
            "default": 0,
            "minimum": 0.0,
            "title": "Seconds",
            "type": "number"
          }
        },
        "required": [
          "kind"
        ],
        "title": "ProjectCoverSelection",
        "type": "object"
      },
      "ProjectCreateRequest": {
        "properties": {
          "description": {
            "default": "",
            "title": "Description",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "useDescriptionForAiPrompt": {
            "default": false,
            "title": "Usedescriptionforaiprompt",
            "type": "boolean"
          }
        },
        "required": [
          "title"
        ],
        "title": "ProjectCreateRequest",
        "type": "object"
      },
      "ProjectListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ProjectSummary"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "ProjectListResponse",
        "type": "object"
      },
      "ProjectPatchRequest": {
        "properties": {
          "cover": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProjectCoverSelection"
              },
              {
                "type": "null"
              }
            ]
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "useDescriptionForAiPrompt": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Usedescriptionforaiprompt"
          }
        },
        "title": "ProjectPatchRequest",
        "type": "object"
      },
      "ProjectRuntimeSummary": {
        "properties": {
          "activeTaskId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Activetaskid"
          },
          "activeTaskTitle": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Activetasktitle"
          },
          "progress": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Progress"
          },
          "promptBatches": {
            "items": {
              "$ref": "#/components/schemas/PromptBatchRuntime"
            },
            "title": "Promptbatches",
            "type": "array"
          },
          "state": {
            "default": "idle",
            "enum": [
              "idle",
              "queued",
              "running",
              "failed"
            ],
            "title": "State",
            "type": "string"
          },
          "videoJobs": {
            "items": {
              "$ref": "#/components/schemas/RuntimeTaskItem"
            },
            "title": "Videojobs",
            "type": "array"
          }
        },
        "title": "ProjectRuntimeSummary",
        "type": "object"
      },
...
      "ProjectSettingsView": {
        "properties": {
          "automaticCoverUrl": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Automaticcoverurl"
          },
          "coverAssetId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Coverassetid"
          },
          "coverUrl": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Coverurl"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "useDescriptionForAiPrompt": {
            "title": "Usedescriptionforaiprompt",
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "title",
          "description",
          "useDescriptionForAiPrompt"
        ],
        "title": "ProjectSettingsView",
        "type": "object"
      },
      "ProjectSummary": {
        "properties": {
          "assetCount": {
            "title": "Assetcount",
            "type": "integer"
          },
          "completedTaskCount": {
            "default": 0,
            "title": "Completedtaskcount",
            "type": "integer"
          },
          "coverUrl": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Coverurl"
          },
          "createdAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Createdat"
          },
          "description": {
            "default": "",
            "title": "Description",
            "type": "string"
          },
          "generationWarnings": {
            "items": {
              "type": "string"
            },
            "title": "Generationwarnings",
            "type": "array"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "newResults": {
            "items": {
              "additionalProperties": {
                "type": "string"
              },
              "type": "object"
            },
            "title": "Newresults",
            "type": "array"
          },
          "status": {
            "enum": [
              "idle",
              "running",
              "completed",
              "failed"
            ],
            "title": "Status",
            "type": "string"
          },
          "taskCount": {
            "title": "Taskcount",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "title": "Updatedat",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "status",
          "taskCount",
          "assetCount",
          "updatedAt"
        ],
        "title": "ProjectSummary",
        "type": "object"
      },
      "ProjectWorkspaceView": {
        "properties": {
          "project": {
            "$ref": "#/components/schemas/WorkspaceProject"
          },
          "runtime": {
            "$ref": "#/components/schemas/ProjectRuntimeSummary"
          },
          "tasks": {
            "items": {
              "$ref": "#/components/schemas/TaskSummary"
            },
            "title": "Tasks",
            "type": "array"
          }
        },
        "required": [
          "project",
          "tasks",
          "runtime"
        ],
        "title": "ProjectWorkspaceView",
        "type": "object"
      },
...
      "PromptBatchRuntime": {
        "properties": {
          "cancelledCount": {
            "title": "Cancelledcount",
            "type": "integer"
          },
          "completedCount": {
            "title": "Completedcount",
            "type": "integer"
          },
          "createdAt": {
            "format": "date-time",
            "title": "Createdat",
            "type": "string"
          },
          "failedCount": {
            "title": "Failedcount",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/RuntimeTaskItem"
            },
            "title": "Items",
            "type": "array"
          },
          "queuedCount": {
            "title": "Queuedcount",
            "type": "integer"
          },
          "runningCount": {
            "title": "Runningcount",
            "type": "integer"
          },
          "state": {
            "title": "State",
            "type": "string"
          }
        },
        "required": [
          "id",
          "createdAt",
          "state",
          "completedCount",
          "failedCount",
          "cancelledCount",
          "queuedCount",
          "runningCount",
          "items"
        ],
        "title": "PromptBatchRuntime",
        "type": "object"
      },
...
      "RuntimeActionRequest": {
        "properties": {
          "action": {
            "enum": [
              "up",
              "down",
              "pause",
              "resume",
              "remove"
            ],
            "title": "Action",
            "type": "string"
          }
        },
        "required": [
          "action"
        ],
        "title": "RuntimeActionRequest",
        "type": "object"
      },
...
      "TaskSummary": {
        "properties": {
          "assetCount": {
            "title": "Assetcount",
            "type": "integer"
          },
          "displayNumber": {
            "title": "Displaynumber",
            "type": "integer"
          },
          "durationSeconds": {
            "title": "Durationseconds",
            "type": "number"
          },
          "generationStatusNote": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Generationstatusnote"
          },
          "generationSummary": {
            "$ref": "#/components/schemas/GenerationSummary"
          },
          "generationWarnings": {
            "items": {
              "type": "string"
            },
            "title": "Generationwarnings",
            "type": "array"
          },
          "hasActivePromptJob": {
            "default": false,
            "title": "Hasactivepromptjob",
            "type": "boolean"
          },
          "hasActiveVideoJob": {
            "default": false,
            "title": "Hasactivevideojob",
            "type": "boolean"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "latestPromptRevisionId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latestpromptrevisionid"
          },
          "latestVideoResultId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latestvideoresultid"
          },
          "previewUrl": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Previewurl"
          },
          "primaryResult": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PrimaryResultView"
              },
              {
                "type": "null"
              }
            ]
          },
          "progress": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Progress"
          },
          "promptEnhancementStatus": {
            "default": "idle",
            "title": "Promptenhancementstatus",
            "type": "string"
          },
          "promptExcerpt": {
            "title": "Promptexcerpt",
            "type": "string"
          },
          "promptReviewStatus": {
            "default": "not_ready",
            "enum": [
              "not_ready",
              "pending_review",
              "approved"
            ],
            "title": "Promptreviewstatus",
            "type": "string"
          },
          "resultCount": {
            "title": "Resultcount",
            "type": "integer"
          },
          "status": {
            "enum": [
              "idle",
              "running",
              "completed",
              "failed"
            ],
            "title": "Status",
            "type": "string"
          },
          "timing": {
            "$ref": "#/components/schemas/TaskTiming"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "videoGenerationStatus": {
            "default": "idle",
            "title": "Videogenerationstatus",
            "type": "string"
          }
        },
        "required": [
          "id",
          "displayNumber",
          "title",
          "promptExcerpt",
          "previewUrl",
          "status",
          "progress",
          "assetCount",
          "resultCount",
          "durationSeconds",
          "generationSummary"
        ],
        "title": "TaskSummary",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "WorkspaceProject": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title"
        ],
        "title": "WorkspaceProject",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "ShotMill",
    "version": "0.3"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/jobs/{job_id}": {
      "get": {
        "operationId": "get_job_api_v1_jobs__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Job",
        "tags": [
          "generation"
        ]
      }
    },
    "/api/v1/projects": {
      "get": {
        "operationId": "list_projects_api_v1_projects_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectListResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Projects",
        "tags": [
          "projects"
        ]
      },
      "post": {
        "operationId": "create_project_api_v1_projects_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Project",
        "tags": [
          "projects"
        ]
      }
    },
...
    "/api/v1/projects/{project_id}": {
      "patch": {
        "operationId": "update_project_api_v1_projects__project_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectPatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Project",
        "tags": [
          "projects"
        ]
      }
    },
...
    "/api/v1/projects/{project_id}/runtime/{kind}/{job_id}": {
      "post": {
        "operationId": "runtime_action_api_v1_projects__project_id__runtime__kind___job_id__post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "kind",
            "required": true,
            "schema": {
              "enum": [
                "prompt",
                "video"
              ],
              "title": "Kind",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RuntimeActionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Runtime Action",
        "tags": [
          "projects"
        ]
      }
    },
    "/api/v1/projects/{project_id}/settings": {
      "get": {
        "operationId": "get_project_settings_api_v1_projects__project_id__settings_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectSettingsView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Project Settings",
        "tags": [
          "projects"
        ]
      }
    },
...
    "/api/v1/projects/{project_id}/tasks/{task_id}/generation": {
      "post": {
        "operationId": "submit_generation_api_v1_projects__project_id__tasks__task_id__generation_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerationSubmitRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Submit Generation",
        "tags": [
          "generation"
        ]
      }
    },
...
    "/api/v1/projects/{project_id}/video-generation-queue/cancel": {
      "post": {
        "operationId": "cancel_queued_videos_api_v1_projects__project_id__video_generation_queue_cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CancelVideoJobsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CancelledVideoJobsView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel Queued Videos",
        "tags": [
          "generation"
        ]
      }
    },
    "/api/v1/projects/{project_id}/workspace": {
      "get": {
        "operationId": "get_project_workspace_api_v1_projects__project_id__workspace_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectWorkspaceView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Project Workspace",
        "tags": [
          "projects"
        ]
      }
    },
...
  }
}